safari-534.53-branch/LayoutTests/platform/mac/accessibility/aria-drag-drop-expected.txt

drop
grab
grab
This tests that the ARIA drag and drop attributes work as intended.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS body.childAtIndex(0).ariaDropEffects is 'copy,move'
PASS body.childAtIndex(1).ariaIsGrabbed is true
PASS body.childAtIndex(2).ariaIsGrabbed is false
PASS body.childAtIndex(1).isAttributeSettable('AXGrabbed') is true
PASS body.childAtIndex(2).isAttributeSettable('AXGrabbed') is true
PASS successfullyParsed is true

TEST COMPLETE
